refactor(harness): rank the counts as counts, not as floats
MidRanks took double[] and the two call sites widened long counts to feed it,
which left the tie test looking like floating point equality — the shape
Sonar's S1244 warns about, and rightly so in general.

Here the exact comparison is the point: two emotes tie when they were used the
same number of times, never when they were used a similar number. Taking long[]
says that in the signature instead of relying on the reader to know it. The
ranks and every figure derived from them are unchanged; small longs widen
losslessly, so this is the same arithmetic with the pointless cast removed.

+++ src/EmotePurge.Worker/Harness/ReplayFidelityCalculator.cs
@@ -574,8 +574,8 @@ public static class ReplayFidelityCalculator
             return null;
         }
 
-        var liveRanks = MidRanks([.. entries.Select(e => (double)e.Live)]);
-        var logRanks = MidRanks([.. entries.Select(e => (double)e.Log)]);
+        var liveRanks = MidRanks([.. entries.Select(e => e.Live)]);
+        var logRanks = MidRanks([.. entries.Select(e => e.Log)]);
 
         double meanLive = liveRanks.Average();
         double meanLog = logRanks.Average();
@@ -594,7 +594,13 @@ public static class ReplayFidelityCalculator
         return liveVariance == 0 || logVariance == 0 ? null : covariance / Math.Sqrt(liveVariance * logVariance);
     }
 
-    private static double[] MidRanks(double[] values)
+    /// <summary>
+    /// Average ranks, with tied values sharing the mean of the ranks they span. Takes the counts as
+    /// they are rather than widened to double: the tie test below is an exact comparison, and on
+    /// counts that is the point — two emotes tie when they were used the same number of times, never
+    /// when they were used a similar number of times.
+    /// </summary>
+    private static double[] MidRanks(long[] values)
     {
         var order = Enumerable.Range(0, values.Length).OrderBy(i => values[i]).ToArray();
         var ranks = new double[values.Length];
